Changed syntax to use braces for scope
This is instead of the previous syntax which used indentation for it, which as we have learnt from Python is not such a good idea!

This change also warranted the use of proper parsing techniques: namely, lexing the input into words by a defined grammar, then parsing those words. So the whole thing was basically rewritten.